files_changed cache command + proper beaver cleanup
Beaver gets dropped and all references to the Arc<Beaver> get dropped
so that reference cunt drops to 0 and the `drop` implementation gets
called on Beaver. This was necesarry to properly write the cache
database, which was necessary to implement files_changed.

files_changed returns true if any of the files given to the function
have changed, or it is called for the first time (or after a clean)